Organize types

+++ llvm/lib/DWARFLinker/Classic/DWARFLinker.cpp
@@ -431,7 +431,7 @@ static void constructSeqOffsettoOrigRowMapping(
     return A.get() < B.get();
   });
 
-  std::vector<size_t> SeqStartRows;
+  std::vector<unsigned> SeqStartRows;
   SeqStartRows.push_back(0);
   for (auto [I, Row] : llvm::enumerate(ArrayRef(LT.Rows).drop_back()))
     if (Row.EndSequence)
@@ -471,7 +471,7 @@ static void constructSeqOffsettoOrigRowMapping(
 
   // Dummy last element to make sure StmtAttrsRef and SeqStartRowsRef always
   // run out first.
-  constexpr size_t DummyKey = UINT64_MAX;
+  constexpr uint64_t DummyKey = UINT64_MAX;
   constexpr unsigned DummyVal = UINT32_MAX;
   LineTableMapping[DummyKey] = DummyVal;
 
@@ -481,7 +481,7 @@ static void constructSeqOffsettoOrigRowMapping(
     auto StmtAttrSmallerThanNext = [N = NextSeqOff](const PatchLocation &SA) {
       return SA.get() < N;
     };
-    auto SeqStartSmallerThanNext = [N = NextRow](const uint64_t &Row) {
+    auto SeqStartSmallerThanNext = [N = NextRow](const unsigned &Row) {
       return Row < N;
     };
     // If both StmtAttrs and SeqStartRows points to value not in
